Add libft files

+++ Projetos/Libft/srcs/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: rickymercury <ricardomedeirosx@gmail.co    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/27 21:46:57 by rickymercur       #+#    #+#             */
-/*   Updated: 2024/10/28 17:16:52 by rickymercur      ###   ########.fr       */
+/*   Updated: 2024/10/30 21:09:11 by rickymercur      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
